# remove universal binaries for ViennaRNA