--- a/recipes/viennarna/2.4.5/build.sh
+++ b/recipes/viennarna/2.4.5/build.sh
@@ -5,7 +5,6 @@
 ##
 if [ `uname` == Darwin ] ; then
     extra_config_options="--disable-openmp 
-                          --enable-universal-binary
                           --without-python3
                           LDFLAGS=-Wl,-headerpad_max_install_names"
 fi