--- a/recipes/viennarna/2.4.5/meta.yaml
+++ b/recipes/viennarna/2.4.5/meta.yaml
@@ -3,7 +3,7 @@ package:
   version: 2.4.5
 
 build:
-  number: 2
+  number: 3
 
 source:
   url: https://www.tbi.univie.ac.at/RNA/download/sourcecode/2_4_x/ViennaRNA-2.4.5.tar.gz