--- a/recipes/viennarna/build.sh
+++ b/recipes/viennarna/build.sh
@@ -5,7 +5,6 @@
 ##
 if [ `uname` == Darwin ] ; then
     extra_config_options="--disable-openmp 
-                          --enable-universal-binary
                           LDFLAGS=-Wl,-headerpad_max_install_names"
 fi