@s22chan 2018-07-23
* the flag in configure causes a 1.33-2.00x slowdown (docs say it
interferes with some link-time optimizations)
* the flag is unnecessary since its intent was to bind with Apple's
stock python distribution (rather than conda)

```
echo 'GCGCUUCGCCGCGCGCC&GCGCUUCGCCGCGCGCA\nGCGCUUCGCCGCGCGCC&GCGCUUCGCCGCGCGCA\nGCGCUUCGCCGCGCGCC&GCGCUUCGCCGCGCGCA\nGCGCUUCGCCGCGCGCC&GCGCUUCGCCGCGCGCA\nGCGCUUCGCCGCGCGCC&GCGCUUCGCCGCGCGCA' | time RNAcofold --noPS -p
```

## before ##
RNAcofold --noPS -p  0.02s user 0.00s system 89% cpu 0.030 total

## after ##
RNAcofold --noPS -p  0.01s user 0.00s system 74% cpu 0.022 tota