@s22chan 2018-07-30
@s-will Hi, that was a smaller repro string, but for 50k of dimers, we can see execution time drop from 2mins to 1min.

LTO is disabled, so I'm unsure why the generated code is faster (icache? other linker optimizations?)