@s-will 2018-07-30
Thanks for the quick response (and of course, many thanks for fixing this---apparently quite significant---issue).
It seems that LTO for ViennaRNA is still deactivated on Mac (due to some other problems) --- maybe @raumzeit has some idea what causes this speedup.