@s22chan 2018-07-30
FYI, enabling LTO (or OpenMP) doesn't seem to make RNAcofold or RNAplex run any faster; not sure about other utilities.